requirements.txt
asgiref==3.7.2
Django==5.0
django-phonenumber-field==7.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
phonenumbers==8.13.26
Pillow==10.1.0
psycopg2==2.9.9
PyJWT==2.8.0
pyotp==2.9.0
python-decouple==3.8
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
